+++ src/main/java/ru/murlov/dto/ExchangeRateDto.java
@@ -2,7 +2,7 @@ package ru.murlov.dto;
 
 import ru.murlov.model.Currency;
 
-public record ExchangeRateDto (Long id,
+public record ExchangeRateDto (int id,
                                Currency base_currency,
                                Currency target_currency,
                                float rate) {
